+++ rotkehlchen/externalapis/coingecko.py
@@ -240,6 +240,7 @@ class Coingecko():
     def __init__(self) -> None:
         self.session = requests.session()
         self.session.headers.update({'User-Agent': 'rotkehlchen'})
+        self.all_coins_cache: Optional[Dict[str, Dict[str, Any]]] = None
 
     @overload
     def _query(
@@ -368,8 +369,22 @@ class Coingecko():
 
         return parsed_data
 
-    def all_coins(self) -> List[Dict[str, Any]]:
-        return self._query(module='coins/list')
+    def all_coins(self) -> Dict[str, Dict[str, Any]]:
+        if self.all_coins_cache is None:
+            response = self._query(module='coins/list')
+            self.all_coins_cache = {}
+            for entry in response:
+                if entry['id'] in self.all_coins_cache:
+                    log.warning(
+                        f'Found duplicate coingecko identifier {entry["id"]} when querying '
+                        f'the list of coingecko assets. Ignoring...',
+                    )
+                    continue
+
+                identifier = entry.pop('id')
+                self.all_coins_cache[identifier] = entry
+
+        return self.all_coins_cache
 
     @staticmethod
     def check_vs_currencies(from_asset: Asset, to_asset: Asset, location: str) -> Optional[str]: